Add OneToMany mapping

+++ src/main/java/com/example/sweater/repos/MessageRepos.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface MessageRepos extends CrudRepository<Message, Integer> {
+public interface MessageRepos extends CrudRepository<Message, Long> {
 
     List<Message> findByTag(String tag);
 }
